fixed assertions in vs2k8

+++ libpdfdoc/src/font/SimpleFont.cc
@@ -367,7 +367,8 @@ void SimpleFont::LoadGlyphs( )
 		wchar_t unicode =
 			m_impl->encoding != 0                                  &&
 			m_impl->face->charmap->encoding != FT_ENCODING_UNICODE ?
-				m_impl->encoding->ToUnicode( char_code ) :
+				m_impl->encoding->ToUnicode(
+					static_cast<unsigned short>(char_code) ) :
 				static_cast<wchar_t>( char_code ) ;
 
 		m_impl->glyphs.insert( std::make_pair(
